Add files via upload
This was a project given to me by one of my dat science teacher and I applies various steps to develop the model and then evaluate that model. the model was for house price prediction . I develop this model by working on data set of 2000 plus rows and columns. I first read the data , understand it and then start cleaning it by help of visualization and then I developed a model for the data from clean data I got. after model creation I evaluate that model. I used python language for my this project.